+++ internal/riverinternaltest/startstoptest/startstoptest.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	"github.com/riverqueue/river/internal/maintenance/startstop"
+	"github.com/riverqueue/river/internal/startstop"
 )
 
 // Stress is a test helper that puts stress on a service's start and stop
